requirements.txt
flask
flask-wtf
wtforms
pytz
requests
paramiko
werkzeug
itsdangerous
python-dotenv
flask-socketio
eventlet